third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-property-text-expected.txt

==== Initial style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testEditProperty
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    content: 'EDITED PROPERTY';
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testBreakingCommentEditProperty
Expected protocol error: SyntaxError The property '/*<--OPENED COMMENT' could not be set.

Running test: testInsertFirstProperty
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    content: 'INSERTED PROPERTY';
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testInsertLastProperty
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
    content: 'INSERTED PROPERTY';
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testInsertMultipleProperties
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    content: 'INSERTED #1';content: 'INSERTED #2';
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testInsertPropertyInEmptyRule
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
    content: 'INSERTED PROPERTY';
}


Running test: testInsertBreakingPropertyInLastEmptyRule
Expected protocol error: SyntaxError The property 'content: 'INSERTED PROPERTY'/*' could not be set.

Running test: testDisableProperty
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    /* margin: 0; */
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testRedo
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
    align-items: center;
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}

==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
    align-items: center;
}

p:first-letter {
    font-size: 200%;
}

.empty-rule {
}


Running test: testInvalidParameters
Expected protocol error: Failed to parse styleId argument
